migrate CardSessions interface into common
