+++ androidApp/src/main/java/com/cards/session/android/MainActivity.kt
@@ -27,6 +27,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.cards.session.cards.sdk.CardSessions
+import com.cards.session.cards.sdk.create
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
