+++ cardsSdk/src/iosMain/kotlin/com/cards/session/cards/sdk/CardSessions.kt
@@ -1,0 +1,5 @@
+package com.cards.session.cards.sdk
+
+actual fun CardSessions.Factory.create(apiKey: String): CardSessions {
+    return CardSessionsImpl.create(apiKey)
+}
