+++ cardsSdk/src/iosMain/kotlin/com/cards/session/cards/sdk/CardSessionsIos.kt
@@ -1,0 +1,7 @@
+package com.cards.session.cards.sdk
+
+object CardSessionsIos {
+    fun create(apiKey: String): CardSessions {
+        return CardSessionsImpl.create(apiKey)
+    }
+}
